qingyang-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai>=1.50.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12
Pillow>=10.0.0
sqlalchemy>=2.0
alembic>=1.13
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.0
